Rename go source files for pfdetect
