+++ go/detectparser/dhcp.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"fmt"
