+++ go/detectparser/fortianalyser_test.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"testing"
