+++ go/detectparser/nexpose.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"regexp"
