+++ go/detectparser/parser.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"context"
