+++ go/detectparser/security_onion_test.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"testing"
