+++ go/detectparser/suricata_md5.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"encoding/json"
