+++ go/detectparser/suricata_md5_test.go
@@ -1,4 +1,4 @@
-package parser
+package detectparser
 
 import (
 	"encoding/json"
